+++ src/tests.rs
@@ -55,6 +55,60 @@ fn test_resampling(
     assert_eq!(resampled, expected);
 }
 
+fn test_resampling_with_none_first(
+    resampling_function: ResamplingFunction<f64, TestSample>,
+    expected: Vec<TestSample>,
+) {
+    let start = DateTime::from_timestamp(0, 0).unwrap();
+    let mut resampler: Resampler<f64, TestSample> =
+        Resampler::new(TimeDelta::seconds(5), resampling_function, 1, start, false);
+    let step = TimeDelta::seconds(1);
+    let data = vec![
+        TestSample::new(start + step, None),
+        TestSample::new(start + step * 2, Some(2.0)),
+        TestSample::new(start + step * 3, Some(3.0)),
+        TestSample::new(start + step * 4, Some(4.0)),
+        TestSample::new(start + step * 5, Some(5.0)),
+        TestSample::new(start + step * 6, None),
+        TestSample::new(start + step * 7, Some(7.0)),
+        TestSample::new(start + step * 8, Some(8.0)),
+        TestSample::new(start + step * 9, Some(9.0)),
+        TestSample::new(start + step * 10, Some(10.0)),
+    ];
+
+    resampler.extend(data);
+
+    let resampled = resampler.resample(start + step * 10);
+    assert_eq!(resampled, expected);
+}
+
+fn test_resampling_with_none_all(
+    resampling_function: ResamplingFunction<f64, TestSample>,
+    expected: Vec<TestSample>,
+) {
+    let start = DateTime::from_timestamp(0, 0).unwrap();
+    let mut resampler: Resampler<f64, TestSample> =
+        Resampler::new(TimeDelta::seconds(5), resampling_function, 1, start, false);
+    let step = TimeDelta::seconds(1);
+    let data = vec![
+        TestSample::new(start + step, None),
+        TestSample::new(start + step * 2, None),
+        TestSample::new(start + step * 3, None),
+        TestSample::new(start + step * 4, None),
+        TestSample::new(start + step * 5, None),
+        TestSample::new(start + step * 6, None),
+        TestSample::new(start + step * 7, None),
+        TestSample::new(start + step * 8, None),
+        TestSample::new(start + step * 9, None),
+        TestSample::new(start + step * 10, None),
+    ];
+
+    resampler.extend(data);
+
+    let resampled = resampler.resample(start + step * 10);
+    assert_eq!(resampled, expected);
+}
+
 #[test]
 fn test_resampling_average() {
     test_resampling(
@@ -62,6 +116,22 @@ fn test_resampling_average() {
         vec![
             TestSample::new(DateTime::from_timestamp(5, 0).unwrap(), Some(3.0)),
             TestSample::new(DateTime::from_timestamp(10, 0).unwrap(), Some(8.0)),
+        ],
+    );
+
+    test_resampling_with_none_first(
+        ResamplingFunction::Average,
+        vec![
+            TestSample::new(DateTime::from_timestamp(5, 0).unwrap(), Some(3.5)),
+            TestSample::new(DateTime::from_timestamp(10, 0).unwrap(), Some(8.5)),
+        ],
+    );
+
+    test_resampling_with_none_all(
+        ResamplingFunction::Average,
+        vec![
+            TestSample::new(DateTime::from_timestamp(5, 0).unwrap(), None),
+            TestSample::new(DateTime::from_timestamp(10, 0).unwrap(), None),
         ],
     );
 }
@@ -75,6 +145,22 @@ fn test_resampling_count() {
             TestSample::new(DateTime::from_timestamp(10, 0).unwrap(), Some(5.0)),
         ],
     );
+
+    test_resampling_with_none_first(
+        ResamplingFunction::Count,
+        vec![
+            TestSample::new(DateTime::from_timestamp(5, 0).unwrap(), Some(4.0)),
+            TestSample::new(DateTime::from_timestamp(10, 0).unwrap(), Some(4.0)),
+        ],
+    );
+
+    test_resampling_with_none_all(
+        ResamplingFunction::Count,
+        vec![
+            TestSample::new(DateTime::from_timestamp(5, 0).unwrap(), Some(0.0)),
+            TestSample::new(DateTime::from_timestamp(10, 0).unwrap(), Some(0.0)),
+        ],
+    );
 }
 
 #[test]
@@ -84,6 +170,22 @@ fn test_resampling_sum() {
         vec![
             TestSample::new(DateTime::from_timestamp(5, 0).unwrap(), Some(15.0)),
             TestSample::new(DateTime::from_timestamp(10, 0).unwrap(), Some(40.0)),
+        ],
+    );
+
+    test_resampling_with_none_first(
+        ResamplingFunction::Sum,
+        vec![
+            TestSample::new(DateTime::from_timestamp(5, 0).unwrap(), Some(14.0)),
+            TestSample::new(DateTime::from_timestamp(10, 0).unwrap(), Some(34.0)),
+        ],
+    );
+
+    test_resampling_with_none_all(
+        ResamplingFunction::Sum,
+        vec![
+            TestSample::new(DateTime::from_timestamp(5, 0).unwrap(), None),
+            TestSample::new(DateTime::from_timestamp(10, 0).unwrap(), None),
         ],
     );
 }
@@ -97,6 +199,22 @@ fn test_resampling_min() {
             TestSample::new(DateTime::from_timestamp(10, 0).unwrap(), Some(6.0)),
         ],
     );
+
+    test_resampling_with_none_first(
+        ResamplingFunction::Min,
+        vec![
+            TestSample::new(DateTime::from_timestamp(5, 0).unwrap(), Some(2.0)),
+            TestSample::new(DateTime::from_timestamp(10, 0).unwrap(), Some(7.0)),
+        ],
+    );
+
+    test_resampling_with_none_all(
+        ResamplingFunction::Min,
+        vec![
+            TestSample::new(DateTime::from_timestamp(5, 0).unwrap(), None),
+            TestSample::new(DateTime::from_timestamp(10, 0).unwrap(), None),
+        ],
+    );
 }
 
 #[test]
@@ -108,6 +226,49 @@ fn test_resampling_max() {
             TestSample::new(DateTime::from_timestamp(10, 0).unwrap(), Some(10.0)),
         ],
     );
+
+    test_resampling_with_none_first(
+        ResamplingFunction::Max,
+        vec![
+            TestSample::new(DateTime::from_timestamp(5, 0).unwrap(), Some(5.0)),
+            TestSample::new(DateTime::from_timestamp(10, 0).unwrap(), Some(10.0)),
+        ],
+    );
+
+    test_resampling_with_none_all(
+        ResamplingFunction::Max,
+        vec![
+            TestSample::new(DateTime::from_timestamp(5, 0).unwrap(), None),
+            TestSample::new(DateTime::from_timestamp(10, 0).unwrap(), None),
+        ],
+    );
+}
+
+#[test]
+fn test_resampling_first() {
+    test_resampling(
+        ResamplingFunction::First,
+        vec![
+            TestSample::new(DateTime::from_timestamp(5, 0).unwrap(), Some(1.0)),
+            TestSample::new(DateTime::from_timestamp(10, 0).unwrap(), Some(6.0)),
+        ],
+    );
+
+    test_resampling_with_none_first(
+        ResamplingFunction::First,
+        vec![
+            TestSample::new(DateTime::from_timestamp(5, 0).unwrap(), None),
+            TestSample::new(DateTime::from_timestamp(10, 0).unwrap(), None),
+        ],
+    );
+
+    test_resampling_with_none_all(
+        ResamplingFunction::First,
+        vec![
+            TestSample::new(DateTime::from_timestamp(5, 0).unwrap(), None),
+            TestSample::new(DateTime::from_timestamp(10, 0).unwrap(), None),
+        ],
+    );
 }
 
 #[test]
@@ -117,6 +278,49 @@ fn test_resampling_last() {
         vec![
             TestSample::new(DateTime::from_timestamp(5, 0).unwrap(), Some(5.0)),
             TestSample::new(DateTime::from_timestamp(10, 0).unwrap(), Some(10.0)),
+        ],
+    );
+
+    test_resampling_with_none_first(
+        ResamplingFunction::Last,
+        vec![
+            TestSample::new(DateTime::from_timestamp(5, 0).unwrap(), Some(5.0)),
+            TestSample::new(DateTime::from_timestamp(10, 0).unwrap(), Some(10.0)),
+        ],
+    );
+
+    test_resampling_with_none_all(
+        ResamplingFunction::Last,
+        vec![
+            TestSample::new(DateTime::from_timestamp(5, 0).unwrap(), None),
+            TestSample::new(DateTime::from_timestamp(10, 0).unwrap(), None),
+        ],
+    );
+}
+
+#[test]
+fn test_resampling_coalesce() {
+    test_resampling(
+        ResamplingFunction::Coalesce,
+        vec![
+            TestSample::new(DateTime::from_timestamp(5, 0).unwrap(), Some(1.0)),
+            TestSample::new(DateTime::from_timestamp(10, 0).unwrap(), Some(6.0)),
+        ],
+    );
+
+    test_resampling_with_none_first(
+        ResamplingFunction::Coalesce,
+        vec![
+            TestSample::new(DateTime::from_timestamp(5, 0).unwrap(), Some(2.0)),
+            TestSample::new(DateTime::from_timestamp(10, 0).unwrap(), Some(7.0)),
+        ],
+    );
+
+    test_resampling_with_none_all(
+        ResamplingFunction::Coalesce,
+        vec![
+            TestSample::new(DateTime::from_timestamp(5, 0).unwrap(), None),
+            TestSample::new(DateTime::from_timestamp(10, 0).unwrap(), None),
         ],
     );
 }